requirements.txt
# requirements.txt - COMPATIBLE VERSIONS
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.2.2
scikit-learn==1.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
langchain>=0.1
langchain-community
langchain-huggingface==0.0.2
langchain-groq==0.1.0
numpy==1.26.4
nltk==3.8.1
tqdm==4.66.1
pypdf==3.17.4
sentence-transformers==2.6.1
langchain-text-splitters>=0.0.1
